requirements.txt
pygame>=2.0.1
numpy>=1.21.2
Pillow>=8.3.2
